ml_service/requirements.txt
lingua-language-detector
fastapi
uvicorn[standard]
pydantic-settings >= 2.12.0
python-dotenv